Correct “published specification” fields (#45)
This is based on feedback by expert & RFC 6838 author Ned Freed. About the field:

"""
Generally speaking it should reference the format specification. However, this
can get tricky when the registration is itself in a standards document that
doesn't itself contain the format specification, but references it. What if the
standards document is updated and the pointer to the proper specification
changes?

In the case of IETF specifications IANA is part of update process and will
update the registration. But that may not be the case with other standards
bodies — they may not see the need to update the specification.

Also remember that a specification is only required for standards tree
types.

tl;dr Reference the actual format specification if possible, but be aware
there can be issues.
"""